The character is unsigned.  Fixes #578

+++ src/modules/rlm_linelog/rlm_linelog.c
@@ -176,7 +176,7 @@ static size_t linelog_escape_func(UNUSED REQUEST *request,
 
 		default:
 			if (outlen <= 4) break;
-			snprintf(out, outlen,  "\\%03o", *in);
+			snprintf(out, outlen,  "\\%03o", (uint8_t) *in);
 			in++;
 			out += 4;
 			outlen -= 4;
